requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptocode==0.1
Django==4.2.10
django-celery-beat==2.6.0
django-timezone-field==6.1.0
djangorestframework==3.15.0
drf-yasg==1.21.7
inflection==0.5.1
kombu==5.3.6
packaging==24.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodomex==3.20.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
